Finalize split to transport crate fix compilation bugs
