+++ Cargo.toml
@@ -31,6 +31,14 @@ log = "0.4.22"
 tokio = { version = "1.43", features = ["full"] }
 maybe-async = "0.2"
 futures-util = { version = "0.3" }
+futures-core = { version = "0.3" }
+tokio-util = { version = "0.7" }
+
+# TLS
+quinn = { version = "0.11.9" }
+reqwest = { version = "0.12", default-features = false }
+rustls = { version = "0.23", default-features = false }
+rustls-platform-verifier = { version = "0.6.1", default-features = false }
 
 # Profiling: release (optimized) with debug info
 [profile.profiling]
