--- crates/smb-direct/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod smbd;
-pub mod rdma;
