+++ crates/smb-transport/src/iovec.rs
