+++ crates/smb/src/connection/config.rs
@@ -1,9 +1,10 @@
 //! Connection configuration settings.
 
-use std::{net::SocketAddr, time::Duration};
+use std::time::Duration;
 
 use smb_dtyp::Guid;
 use smb_msg::Dialect;
+use smb_transport::TransportConfig;
 
 /// Specifies the encryption mode for the connection.
 /// Use this as part of the [ConnectionConfig] to specify the encryption mode for the connection.
@@ -17,31 +18,6 @@ pub enum EncryptionMode {
     /// Encryption is disabled, server might fail the connection if it requires encryption.
     Disabled,
 }
-
-/// Specifies the transport protocol to be used for the connection.
-#[derive(Debug, Default, Clone, PartialEq, Eq)]
-pub enum TransportConfig {
-    /// Use TCP transport protocol.
-    #[default]
-    Tcp,
-    /// Use NetBIOS over TCP transport protocol.
-    NetBios,
-    #[cfg(feature = "quic")]
-    /// Use SMB over QUIC transport protocol.
-    /// Note that this is only suported in dialects 3.1.1 and above.
-    Quic(QuicConfig),
-}
-
-#[cfg(feature = "quic")]
-#[derive(Debug, Default, Clone, PartialEq, Eq)]
-pub struct QuicConfig {
-    pub local_address: Option<SocketAddr>,
-    pub cert_validation: QuicCertValidationOptions,
-}
-
-#[cfg(feature = "rdma")]
-#[derive(Debug, Default, Clone, PartialEq, Eq)]
-pub struct RdmaConfig {}
 
 #[derive(Debug, Default, Clone, PartialEq, Eq)]
 pub struct MultiChannelConfig {
