+++ crates/smb/src/error.rs
@@ -1,5 +1,6 @@
 use std::{num::TryFromIntError, sync::PoisonError};
 
+use smb_transport::TransportError;
 use thiserror::Error;
 
 use crate::{UncPath, connection::TransformError, sync_helpers::AcquireError};
