+++ crates/smb/src/session/encryptor_decryptor.rs
@@ -5,8 +5,9 @@ use rand::RngCore;
 use rand::rngs::OsRng;
 use std::io::Cursor;
 
-use crate::{crypto, util::iovec::IoVec};
+use crate::crypto;
 use smb_msg::{Response, encrypted::*};
+use smb_transport::IoVec;
 
 #[derive(Debug)]
 pub struct MessageEncryptor {
