+++ crates/smb/src/tree.rs
@@ -2,13 +2,9 @@ use std::sync::Arc;
 use std::sync::atomic::{AtomicU32, Ordering};
 
 use maybe_async::*;
+use smb_msg::{FileId, FsctlRequest, IoctlRequest, IoctlRequestFlags};
 
 use crate::connection::connection_info::ConnectionInfo;
-use crate::packets::fscc::FileAttributes;
-use crate::packets::smb2::{
-    CreateOptions, FileId, FsctlRequest, IoctlRequest, IoctlRequestFlags, NetworkInterfaceInfo,
-    QueryNetworkInterfaceInfoRequest, RequestContent, ShareFlags, ShareType,
-};
 use crate::resource::FileCreateArgs;
 use smb_fscc::{FileAccessMask, FileAttributes};
 use smb_msg::{
@@ -24,8 +20,10 @@ use crate::{
     session::SessionMessageHandler,
 };
 mod dfs_tree;
+mod ipc_tree;
 use crate::msg_handler::OutgoingMessage;
 pub use dfs_tree::*;
+pub use ipc_tree::*;
 
 type Upstream = HandlerReference<SessionMessageHandler>;
 
@@ -206,6 +204,15 @@ impl Tree {
         Ok(DfsRootTreeRef::new(self))
     }
 
+    pub fn as_ipc_tree(&self) -> crate::Result<IpcTreeRef<'_>> {
+        let info = self.handler.info()?;
+        if info.share_type != ShareType::Pipe {
+            return Err(Error::InvalidState("Tree is not an IPC tree".to_string()));
+        }
+
+        IpcTreeRef::new(self)
+    }
+
     /// Disconnects from the tree (share) on the server.
     ///
     /// After calling this method, none of the resources held open by the tree are accessible.
@@ -214,32 +221,8 @@ impl Tree {
         self.handler.disconnect().await?;
         Ok(())
     }
-    
-    #[maybe_async]
-    pub async fn query_network_interfaces(&self) -> crate::Result<Vec<NetworkInterfaceInfo>> {
-        if self.handler.connect_info.get().map(|i| i.share_type) != Some(ShareType::Pipe) {
-            return Err(Error::InvalidState(
-                "Network interfaces can only be queried on IPC shares".to_string(),
-            ));
-        }
-        if !self.conn_info.config.multichannel.enabled {
-            // Server might decline + this is irrelevant!
-            return Err(Error::InvalidState(
-                "Network interfaces can only be queried when multi-channel is enabled".to_string(),
-            ));
-        }
 
-        const QUERY_NETOWKR_INTERFACE_MAX_OUTPUT: u32 = 2u32.pow(16);
-        let interface_info = self
-            .fsctl_with_options(
-                QueryNetworkInterfaceInfoRequest(()),
-                QUERY_NETOWKR_INTERFACE_MAX_OUTPUT,
-            )
-            .await?;
-
-        Ok(interface_info.into())
-    }
-
+    // TODO: Make it common with ResourceHandle::fsctl_with_options
     #[maybe_async]
     pub(crate) async fn fsctl_with_options<T: FsctlRequest>(
         &self,
@@ -247,7 +230,8 @@ impl Tree {
         max_output_response: u32,
     ) -> crate::Result<T::Response> {
         const NO_INPUT_IN_RESPONSE: u32 = 0;
-        self.handler
+        let response = self
+            .handler
             .send_recv(RequestContent::Ioctl(IoctlRequest {
                 ctl_code: T::FSCTL_CODE as u32,
                 file_id: FileId::FULL,
@@ -260,7 +244,8 @@ impl Tree {
             .message
             .content
             .to_ioctl()?
-            .parse_fsctl::<T::Response>()
+            .parse_fsctl::<T::Response>()?;
+        Ok(response)
     }
 }
