--- crates/smb/src/util.rs
@@ -1,3 +1,0 @@
-pub mod iovec;
-
-pub use iovec::{IoVec, IoVecBuf};
